+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - improvements to our Bazel BUILD file support
 - added a feature to watch for mobile device availability
 - fix a syntax highlighting issue with `sync`
+- we no long scan for Dart projects when Atom is opened from the user's home directory
 
 ## 0.5.3
 - re-enable the super-mixins analyzer setting